+++ src/Digdir.Domain.Dialogporten.Application/Features/V1/EndUser/DialogSeenLogs/Queries/Search/SeenLogDto.cs
@@ -1,3 +1,5 @@
+using Digdir.Domain.Dialogporten.Application.Features.V1.EndUser.Common.Actors;
+
 namespace Digdir.Domain.Dialogporten.Application.Features.V1.EndUser.DialogSeenLogs.Queries.Search;
 
 public sealed class SeenLogDto
@@ -5,15 +7,8 @@ public sealed class SeenLogDto
     public Guid Id { get; set; }
     public DateTimeOffset SeenAt { get; set; }
 
-    public SeenLogSeenByActorDto SeenBy { get; set; } = null!;
+    public ActorDto SeenBy { get; set; } = null!;
 
     public bool IsViaServiceOwner { get; set; }
     public bool IsCurrentEndUser { get; set; }
-}
-
-public sealed class SeenLogSeenByActorDto
-{
-    public Guid Id { get; set; }
-    public string ActorName { get; set; } = null!;
-    public string ActorId { get; set; } = null!;
 }