+++ src/Digdir.Domain.Dialogporten.Application/Features/V1/ServiceOwner/DialogActivities/Queries/Get/ActivityDto.cs
@@ -1,5 +1,5 @@
 using Digdir.Domain.Dialogporten.Application.Features.V1.Common.Localizations;
-using Digdir.Domain.Dialogporten.Domain.Actors;
+using Digdir.Domain.Dialogporten.Application.Features.V1.ServiceOwner.Common.Actors;
 using Digdir.Domain.Dialogporten.Domain.Dialogs.Entities.Activities;
 
 namespace Digdir.Domain.Dialogporten.Application.Features.V1.ServiceOwner.DialogActivities.Queries.Get;
@@ -16,14 +16,6 @@ public sealed class ActivityDto
 
     public Guid? TransmissionId { get; set; }
 
-    public ActivityPerformedByActorDto PerformedBy { get; set; } = null!;
+    public ActorDto PerformedBy { get; set; } = null!;
     public List<LocalizationDto> Description { get; set; } = [];
-}
-
-public sealed class ActivityPerformedByActorDto
-{
-    public Guid Id { get; set; }
-    public ActorType.Values ActorType { get; set; }
-    public string? ActorName { get; set; }
-    public string? ActorId { get; set; }
 }