+++ src/Digdir.Domain.Dialogporten.Application/Features/V1/ServiceOwner/Dialogs/Queries/Search/MappingProfile.cs
@@ -32,15 +32,8 @@ internal sealed class MappingProfile : Profile
         CreateMap<DialogSeenLog, DialogSeenLogDto>()
             .ForMember(dest => dest.SeenAt, opt => opt.MapFrom(src => src.CreatedAt));
 
-        CreateMap<DialogSeenLogSeenByActor, DialogSeenLogSeenByActorDto>()
-            .ForMember(dest => dest.ActorId, opt => opt.MapFrom(src => src.ActorId));
-
         CreateMap<DialogActivity, DialogActivityDto>()
             .ForMember(dest => dest.Type, opt => opt.MapFrom(src => src.TypeId));
-
-        CreateMap<DialogActivityPerformedByActor, DialogActivityPerformedByActorDto>()
-            .ForMember(dest => dest.ActorType, opt => opt.MapFrom(src => src.ActorTypeId))
-            .ForMember(dest => dest.ActorId, opt => opt.MapFrom(src => src.ActorId));
 
         CreateMap<List<DialogContent>?, ContentDto?>()
             .ConvertUsing<DialogContentOutputConverter<ContentDto>>();