+++ src/Digdir.Domain.Dialogporten.WebApi/Program.cs
@@ -52,6 +52,11 @@ static void BuildAndRun(string[] args, TelemetryConfiguration telemetryConfigura
 {
     var builder = WebApplication.CreateBuilder(args);
 
+    builder.WebHost.ConfigureKestrel(kestrelOptions =>
+    {
+        kestrelOptions.Limits.MaxRequestBodySize = Constants.MaxRequestBodySize;
+    });
+
     builder.Host.UseSerilog((context, services, configuration) => configuration
         .MinimumLevel.Warning()
         .ReadFrom.Configuration(context.Configuration)
@@ -194,6 +199,7 @@ static void BuildAndRun(string[] args, TelemetryConfiguration telemetryConfigura
                 document.Generator = null;
                 document.ReplaceProblemDetailsDescriptions();
                 document.ReplaceRequestExampleBodies();
+                document.MakeCollectionsNullable();
             };
         }, uiConfig =>
         {