+++ tests/k6/tests/serviceowner/performance/create-dialog.js
@@ -1,30 +1,24 @@
-import { postSO, expect, describe } from "../../../common/testimports.js";
-import { SharedArray } from 'k6/data';
-import papaparse from 'https://jslib.k6.io/papaparse/5.1.1/index.js';
-import { default as dialogToInsert } from '../testdata/01-create-dialog.js';
+/**
+ * Performance test for creating a dialog
+ * Run: k6 run tests/k6/tests/serviceowner/performance/create-dialog.js --vus 1 --iterations 1
+ */
 import { randomItem } from 'https://jslib.k6.io/k6-utils/1.2.0/index.js';
-
-const filenameServiceowners = '../../performancetest_data/.serviceowners-with-tokens.csv';
-const filenameEndusers = `../../performancetest_data/endusers-${__ENV.API_ENVIRONMENT}.csv`;
-
-const serviceOwners = new SharedArray('serviceOwners', function () {
-  return papaparse.parse(open(filenameServiceowners), { header: true, skipEmptyLines: true }).data;
-});
-
-const endUsers = new SharedArray('endUsers', function () {
-    return papaparse.parse(open(filenameEndusers), { header: true, skipEmptyLines: true }).data;
-  });
+import { getDefaultThresholds } from '../../performancetest_common/getDefaultThresholds.js';
+import { createDialog } from '../../performancetest_common/createDialog.js';
+import { serviceOwners, endUsers } from '../../performancetest_common/readTestdata.js';
 
 export let options = {
     summaryTrendStats: ['avg', 'min', 'med', 'max', 'p(95)', 'p(99)', 'p(99.5)', 'p(99.9)', 'count'],
-    thresholds: {
-        'http_req_duration{scenario:default}': [`max>=0`],
-        'http_req_duration{name:create dialog}': [],
-        'http_reqs{name:create dialog}': [],
-    },
+    thresholds: getDefaultThresholds(['http_req_duration', 'http_reqs'],['create dialog'])
 };
 
 export default function() {
+    if (!endUsers || endUsers.length === 0) {
+        throw new Error('No end users loaded for testing');
+    }
+    if (!serviceOwners || serviceOwners.length === 0) {
+        throw new Error('No service owners loaded for testing');
+    } 
     if ((options.vus === undefined || options.vus === 1) && (options.iterations === undefined || options.iterations === 1)) {
       createDialog(serviceOwners[0], endUsers[0]);
     }
@@ -33,17 +27,3 @@ export default function() {
     }
   }
 
-export function createDialog(serviceOwner, endUser) {  
-    var paramsWithToken = {
-        headers: {
-            Authorization: "Bearer " + serviceOwner.token
-        },
-        tags: { name: 'create dialog' }
-    }
-
-    describe('create dialog', () => {
-        let r = postSO('dialogs', dialogToInsert(endUser.ssn), paramsWithToken);  
-        expect(r.status, 'response status').to.equal(201);
-    });
-    
-}